Improved the time analytical code by avoiding divions by zero. Still problems though. Kristofer
git-svn-id: http://svn.tracmass.org/trunk@169 2ee5d717-d111-0410-bb1d-d2df4af4126a

git-svn-id: https://github.com/TRACMASS/tracmass.git@139 dd566b57-2d5a-cdae-10d7-ce57b7db7bff

+++ trunk/src/init_seed.f95
@@ -13,7 +13,7 @@ subroutine init_seed()
   LOGICAL                                    :: fileexists
   INTEGER                                    :: i,j,k,ist,jst,kst
   INTEGER                                    :: filestat
-  INTEGER, ALLOCATABLE, DIMENSION(:,:)       :: seedMask
+  INTEGER*4, ALLOCATABLE, DIMENSION(:,:)     :: seedMask
   select case (seedType)
      
   case (1) ! === ist, jst, kst method ===
